refactor(markdown): extract shared types into voicecli.markdown_types
Move Segment, TTSDocument, _INSTRUCT_PARTS, compose_instruct,
parse_frontmatter and strip_markdown into a new dedicated
voicecli.markdown_types module sitting in its own (deeper) layer.

This breaks the eager module-load cycle between voicecli._markdown_directives
and voicecli.markdown: the directives module now depends only on the
primitive types module, eliminating the back-reference. voicecli.markdown
re-exports the types so existing call sites (api, translate, engines, tests)
remain unchanged.

The .importlinter ignore entry voicecli._markdown_directives -> voicecli.markdown
is removed; the lazy markdown -> _markdown_directives delegation remains
covered by its existing peer-import ignore.

Closes #142
